style : main page & teamInfo
style : main page & teamInfo

update frontend main page
update frontend teamInfo
