+++ frontend/src/App.vue
@@ -141,7 +141,9 @@ table {
   border-spacing: 0;
 }
 #app {
-  width: 100%;
-  height: 100%;
+  width: 100vw;
+  height: 100vh;
+  margin: 0;
+  padding: 0;
 }
 </style>
