+++ frontend/src/components/Video.vue
@@ -6,7 +6,6 @@
       width="100%"
       height="100%"
       data-setup="{}"
-      @play="onPlay"
     >
       <source src="@/assets/video.mp4" type="video/mp4" />
     </video>
